--- a/.github/workflows/haskell.yml
+++ b/.github/workflows/haskell.yml
@@ -49,7 +49,7 @@ jobs:
 
     - name: Configure the build
       run: |
-        cabal configure --enable-test --enable-benchmark --ghc-options="-Werror" --ghc-options="-fno-ignore-asserts"
+        cabal configure --enable-tests --enable-benchmark --ghc-options="-Werror" --ghc-options="-fno-ignore-asserts"
         cat cabal.project.local
 
     - name: Record cabal dependencies
@@ -138,7 +138,7 @@ jobs:
           ${{ runner.os }}-${{ matrix.ghc }}-${{ env.cache-name }}-
 
     - name: Install stylish-haskell
-      run: cabal install stylish-haskell --constraint 'stylish-haskell == 0.14.6.0'
+      run: cabal install --ignore-project stylish-haskell --constraint 'stylish-haskell == 0.14.6.0'
 
     - name: Record stylish-haskell version
       run: |