--- a/bloomfilter/Data/BloomFilter/Hash.hs
+++ b/bloomfilter/Data/BloomFilter/Hash.hs
@@ -1,5 +1,6 @@
-{-# LANGUAGE BangPatterns, CPP, ForeignFunctionInterface,
+{-# LANGUAGE BangPatterns, CPP, CApiFFI, ForeignFunctionInterface,
     TypeOperators, RoleAnnotations, MagicHash, UnliftedFFITypes #-}
+{-# LANGUAGE InstanceSigs #-}
 
 -- |
 -- Module: Data.BloomFilter.Hash
@@ -24,125 +25,40 @@ module Data.BloomFilter.Hash
       Hash
     -- * Basic hash functionality
     , Hashable(..)
-    , hash32
     , hash64
-    , hashSalt32
-    , hashSalt64
     -- * Compute a family of hash values
-    , hashes
     , CheapHashes (..)
     , cheapHashes
     , evalCheapHashes
     , makeCheapHashes
-    -- * Hash functions for 'Storable' instances
-    , hashOne32
-    , hashOne64
-    , hashList32
-    , hashList64
-    , alignedHashBA
+    -- * Hash functions
+    , hashByteArray
     ) where
 
-import Control.Monad (foldM)
-import Data.Bits ((.&.), (.|.), unsafeShiftL, unsafeShiftR, xor)
 import Data.Array.Byte (ByteArray (..))
-import GHC.Exts (ByteArray#)
-import Data.List (unfoldr)
-import Data.Int (Int8, Int16, Int32, Int64)
-import Data.Word (Word8, Word16, Word32, Word64)
-import Foreign.C.String (CString)
-#if __GLASGOW_HASKELL__ >= 704
-import Foreign.C.Types (CInt(..), CSize(..))
-#else
-import Foreign.C.Types (CInt, CSize)
-#endif
-import Foreign.ForeignPtr (withForeignPtr)
-import Foreign.Marshal.Alloc (alloca)
-import Foreign.Marshal.Array (allocaArray, withArrayLen)
-import Foreign.Ptr (Ptr, castPtr, nullPtr, plusPtr)
-import Foreign.Storable (Storable, peek, poke, sizeOf)
-import System.IO.Unsafe (unsafePerformIO)
-import Data.ByteString.Internal (ByteString(..))
+import Data.Bits (unsafeShiftR)
+import Data.Word (Word32, Word64)
+import XXH3 (xxh3_64bit_withSeed_ba, xxh3_64bit_withSeed_bs)
+
 import qualified Data.ByteString as SB
-import qualified Data.ByteString.Lazy.Internal as LB
-import qualified Data.ByteString.Lazy as LB
+import qualified Data.Primitive as P
 
 -- | A hash value is 32 bits wide.  This limits the maximum size of a
 -- filter to about four billion elements, or 512 megabytes of memory.
 type Hash = Word32
 
-#include "HsBaseConfig.h"
-
--- Make sure we're not performing any expensive arithmetic operations.
--- import Prelude hiding ((/), (*), div, divMod, mod, rem)
-
-foreign import ccall unsafe "lookup3.h _jenkins_hashword" hashWord
-    :: Ptr Word32 -> CSize -> Word32 -> IO Word32
-
-foreign import ccall unsafe "lookup3.h _jenkins_hashwordOff" hashWordBA
-    :: ByteArray# -> CSize -> CSize -> Word32 -> IO Word32
-
-foreign import ccall unsafe "lookup3.h _jenkins_hashword2" hashWord2
-    :: Ptr Word32 -> CSize -> Ptr Word32 -> Ptr Word32 -> IO ()
-
-foreign import ccall unsafe "lookup3.h _jenkins_hashlittle" hashLittle
-    :: Ptr a -> CSize -> Word32 -> IO Word32
-
-foreign import ccall unsafe "lookup3.h _jenkins_hashlittleOff" hashLittleBA
-    :: ByteArray# -> CSize -> CSize -> Word32 -> IO Word32
-
-foreign import ccall unsafe "lookup3.h _jenkins_hashlittle2" hashLittle2
-    :: Ptr a -> CSize -> Ptr Word32 -> Ptr Word32 -> IO ()
-
 class Hashable a where
-    -- | Compute a 32-bit hash of a value.  The salt value perturbs
-    -- the result.
-    hashIO32 :: a               -- ^ value to hash
-             -> Word32          -- ^ salt
-             -> IO Word32
-
     -- | Compute a 64-bit hash of a value.  The first salt value
     -- perturbs the first element of the result, and the second salt
     -- perturbs the second.
-    hashIO64 :: a               -- ^ value to hash
-             -> Word64           -- ^ salt
-             -> IO Word64
-    hashIO64 v salt = do
-                   let s1 = fromIntegral (salt `unsafeShiftR` 32) .&. maxBound
-                       s2 = fromIntegral salt
-                   h1 <- hashIO32 v s1
-                   h2 <- hashIO32 v s2
-                   return $ (fromIntegral h1 `unsafeShiftL` 32) .|. fromIntegral h2
+    hashSalt64
+        :: Word64               -- ^ salt
+        -> a           -- ^ value to hash
+        -> Word64
 
--- | Compute a 32-bit hash.
-hash32 :: Hashable a => a -> Word32
-hash32 = hashSalt32 0x16fc397c
-
+-- | Compute a 64-bit hash.
 hash64 :: Hashable a => a -> Word64
-hash64 = hashSalt64 0x16fc397cf62f64d3
-
--- | Compute a salted 32-bit hash.
-hashSalt32 :: Hashable a => Word32  -- ^ salt
-           -> a                 -- ^ value to hash
-           -> Word32
-{-# INLINE hashSalt32 #-}
-hashSalt32 salt k = unsafePerformIO $ hashIO32 k salt
-
--- | Compute a salted 64-bit hash.
-hashSalt64 :: Hashable a => Word64  -- ^ salt
-           -> a                 -- ^ value to hash
-           -> Word64
-{-# INLINE hashSalt64 #-}
-hashSalt64 salt k = unsafePerformIO $ hashIO64 k salt
-
--- | Compute a list of 32-bit hashes.  The value to hash may be
--- inspected as many times as there are hashes requested.
-hashes :: Hashable a => Int     -- ^ number of hashes to compute
-       -> a                     -- ^ value to hash
-       -> [Hash]
-hashes n v = unfoldr go (n,0x3f56da2d)
-    where go (k,s) | k <= 0    = Nothing
-                   | otherwise = let s' = hashSalt32 s v
-                                 in Just (s', (k-1,s'))
+hash64 = hashSalt64 0
 
 -- | Compute a list of 32-bit hashes relatively cheaply.  The value to
 -- hash is inspected at most twice, regardless of the number of hashes
@@ -165,19 +81,12 @@ evalCheapHashes (CheapHashes h1 h2) i = h1 + (h2 `unsafeShiftR` i)
 
 makeCheapHashes :: Hashable a => a -> CheapHashes a
 {-# SPECIALIZE makeCheapHashes :: SB.ByteString -> CheapHashes SB.ByteString #-}
-{-# SPECIALIZE makeCheapHashes :: LB.ByteString -> CheapHashes LB.ByteString #-}
-{-# SPECIALIZE makeCheapHashes :: String -> CheapHashes String #-}
-makeCheapHashes v = CheapHashes h1 h2
-    where h1 = fromIntegral (h `unsafeShiftR` 32)
-          h2 = fromIntegral h
-          h = hashSalt64 0x9150a946c4a8966e v
+makeCheapHashes v = CheapHashes (fromIntegral (hashSalt64 0 v)) (fromIntegral (hashSalt64 1 v))
 
 cheapHashes :: Hashable a => Int -- ^ number of hashes to compute
             -> a                 -- ^ value to hash
             -> [Hash]
 {-# SPECIALIZE cheapHashes :: Int -> SB.ByteString -> [Hash] #-}
-{-# SPECIALIZE cheapHashes :: Int -> LB.ByteString -> [Hash] #-}
-{-# SPECIALIZE cheapHashes :: Int -> String -> [Hash] #-}
 cheapHashes k v = go 0
     where !ch = makeCheapHashes v
 
@@ -185,238 +94,13 @@ cheapHashes k v = go 0
           go !i | i == k = []
                 | otherwise = evalCheapHashes ch i : go (i + 1)
 
-instance Hashable () where
-    hashIO32 _ salt = return salt
-
-instance Hashable Integer where
-    hashIO32 k salt | k < 0 = hashIO32 (unfoldr go (-k))
-                                   (salt `xor` 0x3ece731e)
-                  | otherwise = hashIO32 (unfoldr go k) salt
-        where go 0 = Nothing
-              go i = Just (fromIntegral i :: Word32, i `unsafeShiftR` 32)
-
-instance Hashable Bool where
-    hashIO32 = hashOne32
-    hashIO64 = hashOne64
-
-instance Hashable Ordering where
-    hashIO32 = hashIO32 . fromEnum
-    hashIO64 = hashIO64 . fromEnum
-
-instance Hashable Char where
-    hashIO32 = hashOne32
-    hashIO64 = hashOne64
-
-instance Hashable Int where
-    hashIO32 = hashOne32
-    hashIO64 = hashOne64
-
-instance Hashable Float where
-    hashIO32 = hashOne32
-    hashIO64 = hashOne64
-
-instance Hashable Double where
-    hashIO32 = hashOne32
-    hashIO64 = hashOne64
-
-instance Hashable Int8 where
-    hashIO32 = hashOne32
-    hashIO64 = hashOne64
-
-instance Hashable Int16 where
-    hashIO32 = hashOne32
-    hashIO64 = hashOne64
-
-instance Hashable Int32 where
-    hashIO32 = hashOne32
-    hashIO64 = hashOne64
-
-instance Hashable Int64 where
-    hashIO32 = hashOne32
-    hashIO64 = hashOne64
-
-instance Hashable Word8 where
-    hashIO32 = hashOne32
-    hashIO64 = hashOne64
-
-instance Hashable Word16 where
-    hashIO32 = hashOne32
-    hashIO64 = hashOne64
-
-instance Hashable Word32 where
-    hashIO32 = hashOne32
-    hashIO64 = hashOne64
-
-instance Hashable Word64 where
-    hashIO32 = hashOne32
-    hashIO64 = hashOne64
-
--- | A fast unchecked shift.  Nasty, but otherwise GHC 6.8.2 does a
--- test and branch on every shift.
-div4 :: CSize -> CSize
-div4 k = fromIntegral ((fromIntegral k :: HTYPE_SIZE_T) `unsafeShiftR` 2)
-
-alignedHash :: Ptr a -> CSize -> Word32 -> IO Word32
-alignedHash ptr bytes salt
-    | bytes .&. 3 == 0 = hashWord (castPtr ptr) (div4 bytes) salt'
-    | otherwise        = hashLittle ptr bytes salt'
-  where salt' = fromIntegral salt
-
-alignedHashBA :: ByteArray -> Int -> Int -> Word32 -> IO Word32
-alignedHashBA (ByteArray ba#) off len salt
-    | bytes .&. 3 == 0
-    , off .&. 3 == 0   = hashWordBA ba# (div4 (fromIntegral off)) (div4 (fromIntegral bytes)) salt'
-    | otherwise        = hashLittleBA ba# (fromIntegral off) (fromIntegral bytes) salt'
-  where
-    salt' = fromIntegral salt
-    bytes = len
-
--- Inlined from Foreign.Marshal.Utils, for performance reasons.
-with :: Storable a => a -> (Ptr a -> IO b) -> IO b
-with val f  =
-  alloca $ \ptr -> do
-    poke ptr val
-    f ptr
-
-alignedHash2 :: Ptr a -> CSize -> Word64 -> IO Word64
-alignedHash2 ptr bytes salt =
-    with (fromIntegral salt) $ \sp -> do
-      let p1 = castPtr sp
-          p2 = castPtr sp `plusPtr` 4
-      doubleHash ptr bytes p1 p2
-      peek sp
-
-doubleHash :: Ptr a -> CSize -> Ptr Word32 -> Ptr Word32 -> IO ()
-doubleHash ptr bytes p1 p2
-          | bytes .&. 3 == 0 = hashWord2 (castPtr ptr) (div4 bytes) p1 p2
-          | otherwise        = hashLittle2 ptr bytes p1 p2
-
 instance Hashable SB.ByteString where
-    hashIO32 bs salt = unsafeUseAsCStringLen bs $ \ptr len ->
-                       alignedHash ptr (fromIntegral len) salt
+    hashSalt64 salt bs = xxh3_64bit_withSeed_bs bs salt
 
-    {-# INLINE hashIO64 #-}
-    hashIO64 bs salt = unsafeUseAsCStringLen bs $ \ptr len ->
-                       alignedHash2 ptr (fromIntegral len) salt
+-- This instance is for tests only
+instance Hashable Word64 where
+    hashSalt64 salt w = case P.primArrayFromList [w] of
+        P.PrimArray ba# -> xxh3_64bit_withSeed_ba (ByteArray ba#) 0 8 salt
 
-rechunk :: LB.ByteString -> [SB.ByteString]
-rechunk s | LB.null s = []
-          | otherwise = let (pre,suf) = LB.splitAt chunkSize s
-                        in  repack pre : rechunk suf
-    where repack    = SB.concat . LB.toChunks
-          chunkSize = fromIntegral LB.defaultChunkSize
-
-instance Hashable LB.ByteString where
-    hashIO32 bs salt = foldM (flip hashIO32) salt (rechunk bs)
-
-    {-# INLINE hashIO64 #-}
-    hashIO64 = hashChunks
-
-instance Hashable a => Hashable (Maybe a) where
-    hashIO32 Nothing salt = return salt
-    hashIO32 (Just k) salt = hashIO32 k salt
-    hashIO64 Nothing salt = return salt
-    hashIO64 (Just k) salt = hashIO64 k salt
-
-instance (Hashable a, Hashable b) => Hashable (Either a b) where
-    hashIO32 (Left a) salt = hashIO32 a salt
-    hashIO32 (Right b) salt = hashIO32 b (salt + 1)
-    hashIO64 (Left a) salt = hashIO64 a salt
-    hashIO64 (Right b) salt = hashIO64 b (salt + 1)
-
-instance (Hashable a, Hashable b) => Hashable (a, b) where
-    hashIO32 (a,b) salt = hashIO32 a salt >>= hashIO32 b
-    hashIO64 (a,b) salt = hashIO64 a salt >>= hashIO64 b
-
-instance (Hashable a, Hashable b, Hashable c) => Hashable (a, b, c) where
-    hashIO32 (a,b,c) salt = hashIO32 a salt >>= hashIO32 b >>= hashIO32 c
-
-instance (Hashable a, Hashable b, Hashable c, Hashable d) =>
-    Hashable (a, b, c, d) where
-    hashIO32 (a,b,c,d) salt =
-        hashIO32 a salt >>= hashIO32 b >>= hashIO32 c >>= hashIO32 d
-
-instance (Hashable a, Hashable b, Hashable c, Hashable d, Hashable e) =>
-    Hashable (a, b, c, d, e) where
-    hashIO32 (a,b,c,d,e) salt =
-        hashIO32 a salt >>= hashIO32 b >>= hashIO32 c >>= hashIO32 d >>= hashIO32 e
-
-instance Storable a => Hashable [a] where
-    hashIO32 = hashList32
-
-    {-# INLINE hashIO64 #-}
-    hashIO64 = hashList64
-
--- | Compute a 32-bit hash of a 'Storable' instance.
-hashOne32 :: Storable a => a -> Word32 -> IO Word32
-hashOne32 k salt = with k $ \ptr ->
-                 alignedHash ptr (fromIntegral (sizeOf k)) salt
-
--- | Compute a 64-bit hash of a 'Storable' instance.
-hashOne64 :: Storable a => a -> Word64 -> IO Word64
-hashOne64 k salt = with k $ \ptr ->
-                   alignedHash2 ptr (fromIntegral (sizeOf k)) salt
-
--- | Compute a 32-bit hash of a list of 'Storable' instances.
-hashList32 :: Storable a => [a] -> Word32 -> IO Word32
-hashList32 xs salt =
-    withArrayLen xs $ \len ptr ->
-        alignedHash ptr (fromIntegral (len * sizeOf (head xs))) salt
-
--- | Compute a 64-bit hash of a list of 'Storable' instances.
-hashList64 :: Storable a => [a] -> Word64 -> IO Word64
-hashList64 xs salt =
-    withArrayLen xs $ \len ptr ->
-        alignedHash2 ptr (fromIntegral (len * sizeOf (head xs))) salt
-
-unsafeUseAsCStringLen :: SB.ByteString -> (CString -> Int -> IO a) -> IO a
-unsafeUseAsCStringLen (PS fp o l) action =
-    withForeignPtr fp $ \p -> action (p `plusPtr` o) l
-
-type HashState = Ptr Word32
-
-foreign import ccall unsafe "lookup3.h _jenkins_little2_begin" c_begin
-    :: Ptr Word32 -> Ptr Word32 -> HashState -> IO ()
-
-foreign import ccall unsafe "lookup3.h _jenkins_little2_frag" c_frag
-    :: Ptr a -> CSize -> HashState -> CSize -> IO CSize
-
-foreign import ccall unsafe "lookup3.h _jenkins_little2_step" c_step
-    :: Ptr a -> CSize -> HashState -> IO CSize
-
-foreign import ccall unsafe "lookup3.h _jenkins_little2_end" c_end
-    :: CInt -> Ptr Word32 -> Ptr Word32 -> HashState -> IO ()
-
-unsafeAdjustCStringLen :: SB.ByteString -> Int -> (CString -> Int -> IO a)
-                       -> IO a
-unsafeAdjustCStringLen (PS fp o l) d action
-  | d > l     = action nullPtr 0
-  | otherwise = withForeignPtr fp $ \p -> action (p `plusPtr` (o + d)) (l - d)
-
-hashChunks :: LB.ByteString -> Word64 -> IO Word64
-hashChunks s salt = do
-    with (fromIntegral salt) $ \sp -> do
-      let p1 = castPtr sp
-          p2 = castPtr sp `plusPtr` 4
-      allocaArray 3 $ \st -> do
-        let step :: LB.ByteString -> Int -> IO Int
-            step (LB.Chunk x xs) off = do
-              unread <- unsafeAdjustCStringLen x off $ \ptr len ->
-                        c_step ptr (fromIntegral len) st
-              if unread > 0
-                then frag xs unread
-                else step xs 0
-            step _ _ = return 0
-
-            frag :: LB.ByteString -> CSize -> IO Int
-            frag c@(LB.Chunk x xs) stoff = do
-              nstoff <- unsafeUseAsCStringLen x $ \ptr len -> do
-                c_frag ptr (fromIntegral len) st stoff
-              if nstoff == 12
-                then step c (fromIntegral (nstoff - stoff))
-                else frag xs nstoff
-            frag LB.Empty stoff = return (fromIntegral (12 - stoff))
-        c_begin p1 p2 st
-        unread <- step s 0
-        c_end (fromIntegral unread) p1 p2 st
-      peek sp
+hashByteArray :: ByteArray -> Int -> Int -> Word64 -> Word64
+hashByteArray = xxh3_64bit_withSeed_ba