… bloomfilter/cbits/lookup3.c
… bloomfilter/cbits/lookup3.h
… xxhash/xxHash-0.8.2/xxhash.h