Use xxhash instead of Jenkins' lookup3
    
Remove most hashable instances, as we don't use them, and we don't need incremental hashing interface elsewhere either.
    
On windows we bundle the library, as C dependencies on windows are PITA. We could also bundle the library everywhere and use INLINE_ALL, but then we'll probably need to tune the C compiler to use SIMD etc.
    
Remove Hashable instance from UTxOKey, we shouldn't hash original keys, but their serialized form.