@phadej 2024-03-11
Before

```text
Benchmarking baseline cost ... 
(This is the cost of just computing the keys.
Subtract this from all subsequent benchmarks.)
Finished: 4.745373309s

Benchmarking makeCheapHashes cost ... 
(This includes the cost of hashing the keys. Subtract this
from all subsequent benchmarks to isolate non-hash costs.)
Finished: 5.386622812s

Benchmarking elemCheapHashes ... 
(this is the simple one-by-one lookup, not bulk lookup)
Finished: 36.245069077s

Benchmarking elemsMany ... 
(this is the bulk lookup)
Finished: 39.384424915s
```

after

```
Benchmarking baseline cost ... 
(This is the cost of just computing the keys.
Subtract this from all subsequent benchmarks.)
Finished: 4.91925596s

Benchmarking makeCheapHashes cost ... 
(This includes the cost of hashing the keys. Subtract this
from all subsequent benchmarks to isolate non-hash costs.)
Finished: 5.286677202s

Benchmarking elemCheapHashes ... 
(this is the simple one-by-one lookup, not bulk lookup)
Finished: 37.824614154s

Benchmarking elemsMany ... 
(this is the bulk lookup)
Finished: 38.892986392s
```


And this considering that new code calculates two hashes (and just throws the high-32bits for now, as the bloomfilter uses only 32bits for now).