@phadej 2024-03-11
Note: using 128 bit hash would be tricky. AFAIK there are no way to FFI 128bits of result at once (without allocating), but if there is, please tell me.