@dcoutts 2024-03-12
With the updated benchmark (mostly presentation changes) in #118:

before
```
Benchmarking makeCheapHashes ... 
(This includes the cost of hashing the keys, less the cost of computing the keys)
Finished.
Time total:        4.00 seconds
Alloc total:       1.18e11 bytes
Time net:          0.69 seconds
Alloc net:         4.00e8 bytes
Time net per key:  2.76e-8 seconds
Alloc net per key: 16.00 bytes
```
after
```
Benchmarking makeCheapHashes ... 
(This includes the cost of hashing the keys, less the cost of computing the keys)
Finished.
Time total:        3.66 seconds
Alloc total:       1.18e11 bytes
Time net:          0.43 seconds
Alloc net:         -736.00 bytes
Time net per key:  1.73e-8 seconds
Alloc net per key: -2.94e-5 bytes
```
Note that the negative net allocations is presumably due to measurement infelicities, what it really shows is that the alloc per key is zero.

So yes, nice result, the time is reduced, despite using a better and bigger hash.