@dcoutts 2024-03-12
> Note: if we want to make our bloom filter code generally useful at some point, we might want to revive some of these Hashable instances and tests

I'm not totally convinced that there's a great benefit to having a hashable class that iterates over the structure of the value, compare to just serialising and hashing.