@dcoutts 2024-03-13
@phadej says
> the incremental hashing interface should be faster to hash a record of dozen ByteArrays and Word64s, than serializing that to (in worst case lazy bytestring) first and then hashing.

> We don't need such generality, someone might.

@jorisdral says
> I am only arguing for generality to make our bloom filter code more useful to the public, and I don't think that has to be at the expense of the Cardano use case. We control what types we hash in lsm-tree and so we can hash RawBytes only, but bloomfilter doesn't have to have that restriction

Yes, fair enough. You're both right.